chore: FastApi logging전역 설정
